Change music index to long instead of int to support longer pieces

+++ ef-core/src/main/java/eflang/core/MusicSource.java
@@ -3,6 +3,6 @@ package eflang.core;
 import java.util.Iterator;
 
 public interface MusicSource extends Iterator<String> {
-    int getPos();
-    void seek(int pos);
+    long getPos();
+    void seek(long pos);
 }
