+++ vibe/src/main/java/eflang/vibe/StepForwardWorker.java
@@ -13,7 +13,7 @@ import java.util.List;
  * @author Ryan Norris
  *
  */
-public class StepForwardWorker extends SwingWorker<Void, Integer> {
+public class StepForwardWorker extends SwingWorker<Void, Long> {
     private VibeController mController;
 
     public StepForwardWorker(VibeController controller) {
@@ -39,9 +39,9 @@ public class StepForwardWorker extends SwingWorker<Void, Integer> {
     }
 
     @Override
-    protected void process(List<Integer> indices) {
+    protected void process(List<Long> indices) {
         //Get EF command index
-        int currentEFCommand = indices.get(indices.size()-1);
+        long currentEFCommand = indices.get(indices.size()-1);
         //Get EAR command index
         int currentEARLine = 0;
         ArrayList<Integer> earLineStartPositions =
